add more safeguards in install/bash.sh to save time/bandwidth

+++ install/bash.sh
@@ -42,30 +42,45 @@ done
 ## https://github.com/ajeetdsouza/zoxide/issues/694#issuecomment-1946069618
 ##
 
-curl -sS https://raw.githubusercontent.com/ajeetdsouza/zoxide/main/install.sh | bash
+if [[ ! -f "${HOME}/.local/bin/zoxide" ]]; then
+    curl -sS https://raw.githubusercontent.com/ajeetdsouza/zoxide/main/install.sh | bash
+else
+    echo "zoxide is already installed on this system"
+fi
 
 ##
 ## Install programs with wget
 ##
 
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo dpkg -i google-chrome-stable_current_amd64.deb
-sudo apt-get install -f
-rm google-chrome-stable_current_amd64.deb
+if command -v google-chrome > /dev/null 2>&1; then
+    echo "Google Chrome is already installed on this system"
+else
+    wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+    sudo dpkg -i google-chrome-stable_current_amd64.deb
+    sudo apt-get install -f
+    rm google-chrome-stable_current_amd64.deb
+fi
 
 ###
 ### Create symlinks for various config/dotfiles
 ###
 
-ln -s ~/projects/dotfiles/config/.gitconfig ~/.gitconfig
-echo "Symlinked .gitconfig"
+ln -s ~/projects/dotfiles/config/.gitconfig ~/.gitconfig \
+&& echo "Symlinked .gitconfig"
 
-mv ~/.bashrc ~/.bashrc.og.bak
-ln -s ~/projects/dotfiles/config/.bashrc ~/.bashrc
-echo "Symlinked .bashrc"
+if [[ ! -f "${HOME}/.bashrc.og.bak" ]]; then
+    mv ~/.bashrc ~/.bashrc.og.bak
+    echo "Backed up original .bashrc to ~/.bashrc.og.bak"
+fi
+ln -s ~/projects/dotfiles/config/.bashrc ~/.bashrc \
+&& echo "Symlinked .bashrc"
 
-ln -s ~/projects/dotfiles/config/.ruff.toml ~/.ruff.toml
-echo "Symlinked .ruff.toml"
+ln -s ~/projects/dotfiles/config/.ruff.toml ~/.ruff.toml \
+&& echo "Symlinked .ruff.toml"
 
-ln -s ~/projects/dotfiles/pass ~/.password-store
-echo "Symlinked pass/ to ~/.password-store"
+if [ ! -L "${HOME}/.password-store" ]; then
+    ln -s ~/projects/dotfiles/pass ~/.password-store
+    echo "Symlinked pass/ to ~/.password-store"
+else
+    echo "password-store has already been symlinked"
+fi
